add remote control interface
This adds a remote control interface which can be loaded on another device. It uses Pusher.com to handle the realtime parts to avoid a lot of requirements on the server hosting this.

Pusher API keys are defined in `js/pusher.php` which are used by both the PHP and JS side of things. Unfortunately Pusher requires server-side code in order to use client messaging, so I've implemented this in PHP as it seemed to be the easiest option to deploy.